requirements.txt
yfinance
pandas
openpyxl
python-dotenv
requests
openai
pywin32
anthropic